pyproject.toml
[tool.poetry]
name = "schedule"
version = "0.1.0"
description = "bot for parsing schedule from google sheet"
authors = ["AsmodaiP <asmoday.p@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-api-python-client = "^2.37.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.6"
python-telegram-bot = "^13.11"
python-dotenv = "^0.19.2"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
